ch(circleci): add circleci badges
- add circleci badges
- break Search Component into two components
