+++ README.md
@@ -1,6 +1,4 @@
-[![Netlify Status](https://api.netlify.com/api/v1/badges/64c45e71-8b32-4a53-9b5f-7e41258e4f9c/deploy-status)](https://app.netlify.com/sites/emmanuelbeja-youtubesearch/deploys)
-
-[![CircleCI](https://circleci.com/gh/EmmanuelBeja/youtubesearch/tree/main.svg?style=svg)](https://circleci.com/gh/EmmanuelBeja/youtubesearch/tree/main)
+[![Netlify Status](https://api.netlify.com/api/v1/badges/64c45e71-8b32-4a53-9b5f-7e41258e4f9c/deploy-status)](https://app.netlify.com/sites/emmanuelbeja-youtubesearch/deploys)[![CircleCI](https://circleci.com/gh/EmmanuelBeja/youtubesearch/tree/main.svg?style=svg)](https://circleci.com/gh/EmmanuelBeja/youtubesearch/tree/main)[![Maintainability](https://api.codeclimate.com/v1/badges/72a4340a4a9074bcc5b8/maintainability)](https://codeclimate.com/github/EmmanuelBeja/youtubesearch/maintainability)[![Test Coverage](https://api.codeclimate.com/v1/badges/72a4340a4a9074bcc5b8/test_coverage)](https://codeclimate.com/github/EmmanuelBeja/youtubesearch/test_coverage)
 
 ## YouTube Search
 This is a simple application search and display data from the YouTube Data API
